MMC: autotest: Fix up the sdcard mount problem
In order to support test (e.x. flash the kernel image with a simple dd
command), DEV_KERNEL exports all of the partitions to user-space.

But due to the limitation of the android vold support(only 6 partitions are
allowed currently), If export all of the partitions, the sdcard will not be
mounted.

4 partitions have been already used, so, only extra 2 left. The param partition
should be exported as the code branch with !EXPORT_PARTITION_DEVICES, Since the
zimage partition just follows the param partition, so, we can simply replace 7
to 2 to export it.

In order to clean up the source code, the magic number 7 is replaced by an
already defined macro PART_EXTRA_PARTITION, and we define PART_EXTRA_PARTITION
as 2 for DEV_KERNEL.