Panic: Dump task state if PANIC_DUMP_TASKS
The tasks' state may help to track down some weird 'panic'
problems, this allows to dump the tasks' state on panic
if CONFIG_PANIC_DUMP_TASKS=1.

Note, since the tasks' state output is huge, we need to
give the ram console buffer more memory, and similarly,
for the dumping opeartion is time-cost, so, the hardware
watchdog should be kept alive during the dumping.