fix bug: max77665 irq.
sometimes if the micorusb persisting with mhl all the time,
even though when the machine is rebooting.
the falling of INT_PMIC1 may not be detected, and the INT_PMIC1
will always be in low status without process any interrupts of
max77665.

according to the datasheet of 77665, if you don't read the INTsrc
reg, the INT-pin will always be pulled low, so we should use
IRQ_TRIGGER_LOW | IRQ_ONESHOT as irqflags of it, and when these
interrupts are cleared in the irq_service_thread, the pin will be
pull up immediately.

Signed-off-by: Mars <caoziqiang@meizu.com>

merge from:
svn://172.16.9.63/svn_src/kernel/meizu_m040/linux-3.0.39-rtm-mp-dev@9071