misc: muic: fix the adc return incorrect.
when the machin reboot. the ADCen and LowPwr will be set to
1 as default, but the LowPwr will exit when attaching a UID or
set it as 0, in this case if the UID had been attached before
MUIC is reset, so the LowPwr will still persist, and the ADC will
return uncorrect value. so anyway, we should renable the ADC and
LowPwr mode when the muic is probed.

so fix this bug.

Signed-off-by: Mars <caoziqiang@meizu.com>

merge from
svn://172.16.9.63/svn_src/kernel/meizu_m040/linux-3.0.39-rtm-mp-dev@9115