s3c2410_wdt: Print backtrace for all cpus in irq mode
The scheduler may hang but the system may be able to response the
interrupts, so, the irq mode still works in some hang cases.

In order to track such problems, we print the backtrace for all cpus.

Please configure the watchdog mode to irq if you want to use this feature.

svn://172.16.9.63/svn_src/kernel/meizu_m040/linux-3.0.39-rtm-mp-dev@9481